bhakti_shorts_automation/requirements.txt
openai>=1.68.2
python-dotenv>=1.0.1
requests>=2.32.3
edge-tts>=6.1.16
gTTS>=2.5.4
moviepy>=2.1.2
google-api-python-client>=2.166.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
APScheduler>=3.10.4
imageio-ffmpeg>=0.5.1
numpy>=2.0
